app/requirements.txt
streamlit
pandas
minio
altair
requests
psycopg2-binary
